Added URL method in view
Every web application framework has to deal with dynamic routing, and must
provide a way to dynamically call an URL, only knowing its route name.
The View class now provides an url() method, which will generate the correct
URL (with URL prefix) of a route, with parameters filled. It can also get
an URL for static ressource (useful in case of URL prefix).